+++ src/VSimApp.cpp
@@ -36,10 +36,10 @@ VSimApp::VSimApp(MainWindow* window)
 	m_narrative_control = new NarrativeControl(this, m_window);
 	m_narrative_player = new NarrativePlayer(this, m_window, m_narrative_control);
 
-	// This really shouldn't be here...
+	// This is a really awkward place... but this has to be done after setting model
 	m_window->outliner()->setModel(&m_model_table_model);
 	m_window->outliner()->header()->resizeSection(0, 200);
-	m_window->outliner()->setGeometry(100, 100, 505, 600);
+	m_window->outliner()->resize(505, 600);
 
 	connect(window, &MainWindow::sOpenFile, this, &VSimApp::openVSim);
 	connect(window, &MainWindow::sSaveFile, this, &VSimApp::saveVSim);